requirements.txt
loguru
matplotlib
numpy
pandas
PyYAML
scipy
setuptools
sounddevice
soundfile
torch
torchaudio
tqdm
